Add files via upload
Check ze comments?
Also, I re-wrote the captcha thing in php so it would show an image with random characters instead of a 4 digit random number. 'Tis not working on my Apache thing. No idea if it's my configuration issue or the code.
Also also, NO idea where to put the captcha_verif.php (All i know is that is has to execute once the submit button is pressed)